lambda-app/cut/requirements.txt
parliament-functions==0.1.0
opencv-python-headless
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
